+++ src/js/packages/@reactpy/client/src/index.ts
@@ -1,5 +1,8 @@
+export * from "./client";
 export * from "./components";
-export * from "./messages";
 export * from "./mount";
-export * from "./reactpy-client";
-export * from "./reactpy-vdom";
+export * from "./types";
+export * from "./vdom";
+export * from "./websocket";
+export { default as React } from "preact/compat";
+export { default as ReactDOM } from "preact/compat";