+++ src/reactpy/__init__.py
@@ -1,5 +1,7 @@
-from reactpy import backend, config, html, logging, sample, svg, types, web, widgets
-from reactpy.backend.utils import run
+from reactpy import asgi, config, logging, types, web, widgets
+from reactpy._html import html
+from reactpy.asgi.middleware import ReactPyMiddleware
+from reactpy.asgi.standalone import ReactPy
 from reactpy.core import hooks
 from reactpy.core.component import component
 from reactpy.core.events import event
@@ -22,12 +24,14 @@ from reactpy.core.vdom import vdom
 from reactpy.utils import Ref, html_to_vdom, vdom_to_html
 
 __author__ = "The Reactive Python Team"
-__version__ = "1.1.0"
+__version__ = "2.0.0a0"
 
 __all__ = [
     "Layout",
+    "ReactPy",
+    "ReactPyMiddleware",
     "Ref",
-    "backend",
+    "asgi",
     "component",
     "config",
     "create_context",
@@ -36,9 +40,6 @@ __all__ = [
     "html",
     "html_to_vdom",
     "logging",
-    "run",
-    "sample",
-    "svg",
     "types",
     "use_callback",
     "use_connection",