+++ tests/test_testing.py
@@ -4,11 +4,10 @@ import os
 import pytest
 
 from reactpy import Ref, component, html, testing
-from reactpy.backend import starlette as starlette_implementation
 from reactpy.logging import ROOT_LOGGER
-from reactpy.sample import SampleApp
 from reactpy.testing.backend import _hotswap
 from reactpy.testing.display import DisplayFixture
+from tests.sample import SampleApp
 
 
 def test_assert_reactpy_logged_does_not_suppress_errors():
@@ -144,19 +143,6 @@ async def test_simple_display_fixture():
         await display.page.wait_for_selector("#sample")
 
 
-def test_if_app_is_given_implementation_must_be_too():
-    with pytest.raises(
-        ValueError,
-        match=r"If an application instance its corresponding server implementation must be provided too",
-    ):
-        testing.BackendFixture(app=starlette_implementation.create_development_app())
-
-    testing.BackendFixture(
-        app=starlette_implementation.create_development_app(),
-        implementation=starlette_implementation,
-    )
-
-
 def test_list_logged_excptions():
     the_error = None
     with testing.capture_reactpy_logs() as records:
@@ -173,7 +159,7 @@ def test_list_logged_excptions():
         assert logged_errors == [the_error]
 
 
-async def test_hostwap_update_on_change(display: DisplayFixture):
+async def test_hotswap_update_on_change(display: DisplayFixture):
     """Ensure shared hotswapping works
 
     This basically means that previously rendered views of a hotswap component get updated
@@ -183,34 +169,39 @@ async def test_hostwap_update_on_change(display: DisplayFixture):
     hotswap component to be updated
     """
 
-    def make_next_count_constructor(count):
-        """We need to construct a new function so they're different when we set_state"""
+    def hotswap_1():
+        return html.div({"id": "hotswap-1"}, 1)
 
-        def constructor():
-            count.current += 1
-            return html.div({"id": f"hotswap-{count.current}"}, count.current)
+    def hotswap_2():
+        return html.div({"id": "hotswap-2"}, 2)
 
-        return constructor
+    def hotswap_3():
+        return html.div({"id": "hotswap-3"}, 3)
 
     @component
     def ButtonSwapsDivs():
         count = Ref(0)
+        mount, hostswap = _hotswap(update_on_change=True)
 
         async def on_click(event):
-            mount(make_next_count_constructor(count))
+            count.set_current(count.current + 1)
+            if count.current == 1:
+                mount(hotswap_1)
+            if count.current == 2:
+                mount(hotswap_2)
+            if count.current == 3:
+                mount(hotswap_3)
 
-        incr = html.button({"on_click": on_click, "id": "incr-button"}, "incr")
-
-        mount, make_hostswap = _hotswap(update_on_change=True)
-        mount(make_next_count_constructor(count))
-        hotswap_view = make_hostswap()
-
-        return html.div(incr, hotswap_view)
+        return html.div(
+            html.button({"on_click": on_click, "id": "incr-button"}, "incr"),
+            hostswap(),
+        )
 
     await display.show(ButtonSwapsDivs)
 
     client_incr_button = await display.page.wait_for_selector("#incr-button")
 
+    await client_incr_button.click()
     await display.page.wait_for_selector("#hotswap-1")
     await client_incr_button.click()
     await display.page.wait_for_selector("#hotswap-2")